Changed alpha to float to make tf less angry

+++ src/sac/sac_agent.py
@@ -101,7 +101,8 @@ class SACAgent():
         self.gradient_steps = gradient_steps
 
         # Setup entropy parameters
-        self.log_alpha = tf.Variable(tf.math.log(alpha), dtype=tf.float32)
+        self.log_alpha = tf.Variable(tf.math.log(tf.cast(alpha, tf.float32)),
+                                     dtype=tf.float32)
         self.alpha = tf.exp(self.log_alpha)
         self.target_entropy = -np.prod(self.env.action_space.shape)
